refactor(community): split community_activity_card.dart into focused modules
Split the 988-line community_activity_card.dart into five focused files
organized by visual concern (closes #429):

- community_activity_card.dart (333): entry widget, chrome, loading/error
- community_activity_bodies.dart (169): card + summary body variants
- community_activity_stats.dart (215): today-stats, simple-count, learners row
- community_activity_avatars.dart (230): avatar widgets + initials utility
- community_activity_metrics.dart (104): inline metric + stat block

Extracted widgets are package-visible with doc comments marking them as
internal building blocks. CommunityActivityCard and
CommunityActivityCardVariant remain the only externally-imported symbols;
home_screen.dart is unchanged.
